add user to workout - schema - refactor - bigger buttons
