+++ app/models/workout.rb
@@ -1,5 +1,12 @@
 class Workout < ActiveRecord::Base
-  scope :latest, -> {
-    order(created_at: :desc)
-  }
+  scope :latest,
+    -> {
+      order(created_at: :desc)
+    }
+
+  scope :for_user,
+    -> (user_id) {
+      where(user_id: user_id)
+    }
+
 end
