+++ app/views/workouts/edit.html.erb
@@ -2,5 +2,5 @@
 
 <%= render 'form' %>
 
-<%= link_to 'Show', @workout %> |
-<%= link_to 'Back', workouts_path %>
+<%= link_to 'Show', @workout, class: 'btn btn-lg btn-info' %> |
+<%= link_to 'Back', workouts_path, class: 'btn btn-lg btn-danger' %>
