+++ app/views/workouts/index.html.erb
@@ -1,7 +1,7 @@
 <p id="notice"><%= notice %></p>
 
-<h1>Listing Workouts</h1>
-<%= link_to 'New Workout', new_workout_path, class: 'btn btn-lg' %>
+<h1>Workouts</h1>
+<%= link_to 'New Workout', new_workout_path, class: 'btn btn-lg btn-success' %>
 
 <table class="table table-condensed table-bordered table-striped">
   <thead>
