+++ app/views/workouts/show.html.erb
@@ -7,5 +7,5 @@
   </p>
 <% end %>
 
-<%= link_to 'Edit', edit_workout_path(@workout) %> |
-<%= link_to 'Back', workouts_path %>
+<%= link_to 'Edit', edit_workout_path(@workout), class: 'btn btn-lg btn-info' %> |
+<%= link_to 'Back', workouts_path, class: 'btn btn-lg btn-danger'%>
