+++ db/migrate/20160515182417_add_user_to_workout.rb
@@ -1,0 +1,5 @@
+class AddUserToWorkout < ActiveRecord::Migration
+  def change
+    add_reference :workouts, :user, index: true, foreign_key: true
+  end
+end
